+++ src/vs/platform/storage/common/storageService.ts
