+++ src/vs/workbench/services/storage/electron-sandbox/storageService.ts
@@ -5,7 +5,7 @@
 
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 import { IMainProcessService } from 'vs/platform/ipc/common/mainProcessService';
-import { NativeStorageService } from 'vs/platform/storage/electron-sandbox/storageService';
+import { NativeStorageService } from 'vs/platform/storage/common/storageService';
 import { IUserDataProfilesService } from 'vs/platform/userDataProfile/common/userDataProfile';
 import { IAnyWorkspaceIdentifier } from 'vs/platform/workspace/common/workspace';
 import { IUserDataProfileService } from 'vs/workbench/services/userDataProfile/common/userDataProfile';
